mvp_bot/requirements.txt
python-telegram-bot==21.6
google-generativeai==0.8.3
python-dotenv==1.0.1
pymupdf==1.24.13
